api/requirements.txt
fastapi
uvicorn[standard]
tortoise-orm[asyncpg]
pydantic[email]
pydantic-settings
python-multipart
python-jose[cryptography]
passlib[bcrypt]
